Module http_parser::parser

Structs

Body
The Body state.

Header
The Header state.

HttpRequestParser
The Parser structure.

Method
The method of the request line.

Request
The HTTP request structure.

RequestLine
The RequestLine, the parser starting state.

Uri
The URI of the request line.

Version
The HTTP version part of the request line.

Enums

ParsingError
Errors types for the parser.

Traits

Parse
The provides the means of state transition for the parser, it provides a single function parse, when called it is supposed to parse the stream until the completion of the current state.

RequestLineParserState
A trait for request line parser states.

RequestParserState
A trait for the request parser states.

Functions

is_crlf
Check if a pair of bytes are CRLF.

is_valid_method
Checks if the given string slice is a valid HTTP method according to IETF RFC 2616 5.1.1.

is_valid_version
Checks if the HTTP version is a valid version.

is_whitespace
Check if a byte is whitespace.

Type Definitions

RequestLineParser
Short form for HttpRequestParser<'a, RequestLine<S>>.

Result
Possible transitions which can error return Result<T, ParsingError>.