Trait http_parser::parser::Parse

pub trait Parse {
    type NextState;
    fn parse(self) -> Self::NextState;
}

The provides the means of state transition for the parser, it provides a single function parse, when called it is supposed to parse the stream until the completion of the current state.

Associated Types

type NextState

NextState type are of kind Parser<'a, State> Sadly we can't do type NextParser = Parser<'a, Self::NextState> and allow the final user to simply define type NextState until https://github.com/rust-lang/rust/issues/29661 is resolved.

Loading content...
Required methods

fn parse(self) -> Self::NextState

Parse the existing content consuming it in the process, in the end, return the next parser state.

Loading content...